Rename thumbnail test to media_msg so that it can cover blurhash too
