+++ native/test/src/tests.rs
@@ -4,6 +4,7 @@ mod calendar;
 mod categories;
 mod formatted_body;
 mod invitation;
+mod media_msg;
 mod msg_edit;
 mod news;
 mod onboarding;
@@ -18,6 +19,5 @@ mod spaces;
 mod super_invites;
 mod tasks;
 mod templates;
-mod thumbnail;
 mod typing;
 mod verification;
